feat: implemented giveaway fix system for endless giveaways

+++ src/Asena.ts
@@ -1,5 +1,6 @@
 import SuperClient from './SuperClient';
 import Constants from './Constants';
+import { validateRaffleText } from './utils/Utils';
 
 export default class Asena extends SuperClient{
 
@@ -32,18 +33,11 @@ export default class Asena extends SuperClient{
         const emojiLength = Constants.CONFETTI_EMOJI.length
         this.on('messageDelete', async message => {
             if(message.author?.id === this.user.id){
-                if(message.content){
-                    const content = message.content.trim()
-                    if(
-                        content.length >= emojiLength * 2 &&
-                        content.substr(0, emojiLength) == Constants.CONFETTI_EMOJI &&
-                        content.substr(content.length - emojiLength, emojiLength) == Constants.CONFETTI_EMOJI
-                    ){
-                        const server = await this.servers.get(message.guild?.id)
-                        const raffle = await server.raffles.get(message.id)
-                        if(raffle && raffle.isContinues()){
-                            await raffle.delete()
-                        }
+                if(message.content && validateRaffleText(message.content)){
+                    const server = await this.servers.get(message.guild?.id)
+                    const raffle = await server.raffles.get(message.id)
+                    if(raffle && raffle.isContinues()){
+                        await raffle.delete()
                     }
                 }
             }
